[refactor] Move search providers to dedicated providers subdirectory
