+++ src/services/gateway/registrySearchGateway.ts
@@ -1,5 +1,5 @@
-import { useAlgoliaSearchProvider } from '@/services/algoliaSearchProvider'
-import { useComfyRegistrySearchProvider } from '@/services/registrySearchProvider'
+import { useAlgoliaSearchProvider } from '@/services/providers/algoliaSearchProvider'
+import { useComfyRegistrySearchProvider } from '@/services/providers/registrySearchProvider'
 import type { SearchNodePacksParams } from '@/types/algoliaTypes'
 import type { components } from '@/types/comfyRegistryTypes'
 import type {
