+++ src/services/providers/algoliaSearchProvider.ts
