+++ src/services/providers/registrySearchProvider.ts
